💻 REFACTOR: new folder structure for testing
