+++ .gitignore
@@ -13,3 +13,4 @@
 .env
 # Dependency directories (remove the comment below to include it)
 # vendor/
+/worki-back
