+++ handler/gig.go
@@ -1,10 +1,9 @@
-package gig
+package handler
 
 import (
 	"net/http"
 	"strconv"
 
-	"github.com/LeoBorquez/worki-back/handler/user"
 	"github.com/LeoBorquez/worki-back/model"
 	"github.com/biezhi/gorm-paginator/pagination"
 	"github.com/labstack/echo"
@@ -14,7 +13,7 @@ import (
 func (h *Handler) CreateGig(c echo.Context) (err error) {
 
 	db := h.DB
-	userID := user.UserIDFromToken(c)
+	userID := UserIDFromToken(c)
 
 	g := &model.Gig{}
 	u := &model.User{}
@@ -45,7 +44,7 @@ func (h *Handler) CreateGig(c echo.Context) (err error) {
 
 // FetchGig return the last gigs added
 func (h *Handler) FetchGig(c echo.Context) (err error) {
-	userID := user.userIDFromToken(c)
+	userID := UserIDFromToken(c)
 	if userID == 0 {
 		return &echo.HTTPError{Code: http.StatusUnauthorized, Message: "Login requested"}
 	}
