+++ handler/test/user_test.go
@@ -1,4 +1,4 @@
-package user
+package test
 
 import (
 	"database/sql"
