+++ server.go
@@ -46,7 +46,7 @@ func main() {
 		},
 	}))
 
-	e.POST("/signup", h.User)
+	e.POST("/signup", h.Signup)
 	e.POST("/login", h.Login)
 	e.POST("/gigs", h.CreateGig)
 	e.GET("/feed", h.FetchGig)
